requirements.txt
# 안정화된 의존성 - 헥토 AI 자동차 분류 V2
# Python 3.8+ 호환성 보장

# 핵심 딥러닝 프레임워크 (버전 범위 안정화)
torch>=2.2.0,<2.7.0
torchvision>=0.21.0,<0.23.0
timm>=0.6.12,<1.0.0

# 데이터 처리 및 증강
albumentations>=1.3.0,<1.4.0
opencv-python>=4.6.0,<4.10.0
# scikit-learn>=1.1.0,<1.5.0
pandas>=1.5.0,<2.3.0
# numpy>=1.21.0,<2.1.0
pillow>=9.0.0,<11.0.0

# 실험 추적 (선택적)
wandb>=0.15.0,<0.17.0
tensorboard>=2.10.0,<2.16.0

# 유틸리티
matplotlib>=3.5.0,<3.9.0
seaborn>=0.11.0,<0.14.0
tqdm>=4.64.0,<5.0.0
omegaconf>=2.3.0,<2.4.0
pyyaml>=6.0.0,<7.0.0

# 개발 도구 (옵션)
pytest>=7.0.0,<8.0.0
black>=22.0.0,<24.0.0
isort>=5.10.0,<6.0.0